feat: merge dual session systems into unified Session with Option<TeeHandle>
- Rewrite session/manager.rs: unified Session struct with TeeHandle behind
  Arc<RwLock<Option<TeeHandle>>> for safe TEE upgrades
- SessionManager now handles both regular and TEE session lifecycles
  (init_tee, shutdown_tee, upgrade_to_tee, process_in_tee)
- Remove TeeManager and TeeSession (deleted tee/manager.rs)
- Simplify SessionRouter: remove tee_manager field, use unified session_manager
- Simplify Gateway: remove tee_manager field, single session_manager for all ops
- Update handler.rs: uses_tee is now async method, not field
- Update websocket.rs: use session_manager.process_in_tee() instead of tee_manager
- Add manual Debug impl for TeeClient
- 243 tests pass, clippy clean

+++ src/privacy/policy.rs
@@ -101,7 +101,7 @@ impl PolicyEngine {
                 }
             }
             SensitivityLevel::Sensitive => {
-                if policy.tee_threshold as u8 <= SensitivityLevel::Sensitive as u8 {
+                if policy.tee_threshold <= SensitivityLevel::Sensitive {
                     PolicyDecision::ProcessInTee
                 } else {
                     PolicyDecision::ProcessLocal
